+++ index_investment/old/Portfolios/fornybarenergiindex.csv
