+++ requirements.txt
@@ -1,5 +1,5 @@
 pandas~=1.2.1
-numpy~=1.19.5
+numpy~=1.23.0
 requests~=2.25.1
 beautifulsoup4~=4.9.3
 selenium~=3.141.0
